+++ pyhidra/uninstall_desktop.py
@@ -7,6 +7,8 @@ if __name__ == "__main__":
         from pyhidra.win_shortcut import remove_shortcut
     elif sys.platform == "linux":
         from pyhidra.linux_shortcut import remove_shortcut
+    elif sys.platform == "darwin":
+        from pyhidra.mac_shortcut import remove_shortcut
     else:
         sys.exit("Unsupported platform")